vm1/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.2.1
flask==3.0.3
Flask-SocketIO==5.3.6
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
llama-cpp-python==0.2.76
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector-python==8.4.0
numpy==1.24.4
openai==1.30.5
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.3
pyarrow==16.0.0
pydantic==2.7.2
pydantic-core==2.18.3
pydeck==0.9.1
pygments==2.18.0
python-dateutil==2.9.0.post0
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.36.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
websocket-client==1.8.0
werkzeug==3.0.3
wsproto==1.2.0
zipp==3.18.1
